# Domain/ValueObjects/ChaveNota.cs
namespace Tax_Document_Processor.Domain.ValueObjects
{
    public class ChaveNota
    {
        public string Value { get; private set; }

        public bool IsNfse => Value.Length == 50;

        public ChaveNota(string value)
        {
            if (value.Length != 44 && value.Length != 50)
                throw new Exception("Chave deve ter 44 ou 50 dígitos");

            Value = value;
        }
    }
}